Solucionando bugs
