--- src/app/components/PostSend/components/interfaces/FormValues.interface.tsx
@@ -1,5 +1,0 @@
-export interface FormValues {
-  description: string;
-  file: File[] | null;
-  userId:string;
-}
