--- src/app/components/PostSend/components/validationSchema.tsx
@@ -1,7 +1,0 @@
-import * as Yup from 'yup';
-const validationSchema = Yup.object().shape({
-  description: Yup.string().required('La descripción es requerida'),
-  userId: Yup.string().required('La descripción es requerida'),
-  file: Yup.mixed<File[]>().nullable(),
-});
-export default validationSchema;
