--- src/app/components/posts/components/interfaces/Comments.interface.tsx
@@ -1,10 +1,0 @@
-export default interface CommentsProps {
-  firstName: string;
-  datePublication: string;
-  description: string;
-  imagen: string | null;
-  id: string;
-  userId: string; 
-  likes: number;
-  comments: [];
-}
