--- src/app/components/posts/components/interfaces/CommentsSend.interface.tsx
@@ -1,5 +1,0 @@
-export default interface CommentsSendProps {
-  id: string;
-  userId: string; 
-  
-}
