--- src/app/components/posts/components/interfaces/Likes.interface.tsx
@@ -1,4 +1,0 @@
-export default interface LikesProps {
-  id: string;
-  userId: string;
-}
