--- src/app/components/posts/components/interfaces/comment.interface.tsx
@@ -1,10 +1,0 @@
-interface Comment {
-  id: string;
-  description: string;
-  firstName: string;
-}
-
-export default interface CommentsPost {
-  comment: Comment[];
- 
-}
